fix: move campfirecontent to src, create empty /app campfire page for next.js
Next does not allow a page to have custom props. Therefore moved campfire with content into src/components, and created a new page inside /app/campfire that displays this content. Updated imports.
