+++ frontend/app/campfire/page.tsx
@@ -1,21 +1,9 @@
-'use client';
+import CampfirePage from '@/src/components/campfire/CampfirePage';
 
-import Campfire from '@/src/components/Campfire';
-import Link from 'next/link';
+export const metadata = {
+  title: 'Campfire',
+};
 
-interface CampfirePageProps {
-  message?: string;
-}
-
-export default function CampfirePage({
-  message = 'Og tent et bål for å holde deg varm',
-}: CampfirePageProps) {
-  return (
-    <div className="flex flex-col items-center justify-center min-h-[60vh]">
-      <Campfire message={message} />
-      <Link href="/" className="mt-4 underline opacity-80 hover:opacity-100">
-        Gå til hjemsiden
-      </Link>
-    </div>
-  );
+export default function CampfireRoute() {
+  return <CampfirePage />;
 }
