+++ frontend/app/not-found.tsx
@@ -1,4 +1,4 @@
-import CampfirePage from './campfire/page';
+import CampfirePage from '../src/components/campfire/CampfirePage';
 
 export default function NotFound() {
   return <CampfirePage message="Denne siden finnes ikke." />;
