+++ frontend/app/offline/page.tsx
@@ -1,4 +1,4 @@
-import CampfirePage from '../campfire/page';
+import CampfirePage from '../../src/components/campfire/CampfirePage';
 
 export const metadata = {
   title: 'Du er offline',
